requirements.txt
langgraph>=0.2.42
langchain-aws>=0.2.10
bedrock-agentcore-starter-toolkit>=0.0.30
boto3>=1.34
python-dotenv>=1.0.1
requests>=2.32.3
boto3
streamlit
mcp
langchain_mcp_adapters